backend/requirements.txt
fastapi
uvicorn
groq
faiss-cpu
sentence-transformers
pymupdf
numpy
